自定义数字键盘、随机键盘完成、Example也完成

+++ WYAKit/Classes/WYAUIKit/WYAShareView/WYAShareView.m
@@ -99,7 +99,7 @@
     if (self.delegate && [self.delegate respondsToSelector:@selector(wya_shareView:didSelectItemAtIndexPath:itemTitle:)]) {
         [self.delegate wya_shareView:self didSelectItemAtIndexPath:indexPath itemTitle:title];
     }
-    NSLog(@"indexPath.row%ld-------title%@",indexPath.row,title);
+    NSLog(@"indexPath.row%ld-------title%@",(long)indexPath.row,title);
 }
 #pragma mark ======= getter
 - (UICollectionView *)collectionView{
